+++ lib/ui/widgets/post_content/widgets/post_poll_content/widgets/post_poll_result_item/index.dart
@@ -1,58 +1,103 @@
 import 'package:flutter/material.dart';
 import 'package:mooncake/entities/entities.dart';
 import 'package:mooncake/ui/theme/spaces.dart';
+import '../../utils.dart' show pollColors;
 
 /// Represents a single row showing the percentage of voting that the
 /// given poll option had.
 class PostPollResultItem extends StatelessWidget {
   final PostPoll poll;
   final PollOption option;
-
-  /// Represents the percentage of times that this option has
-  /// been chosen over all the others.
-  final double percentage;
+  final int index;
+  final int currentPollLength;
 
   PostPollResultItem({
     Key key,
     @required this.poll,
     @required this.option,
-  })  : percentage = (poll.userAnswers
-                .where((answer) => answer.answer == option.id)
-                .length) /
-            poll.userAnswers.length,
+    @required this.index,
+  })  : currentPollLength = poll.userAnswers
+            .where((answer) => answer.answer == option.id)
+            .length,
         super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    final double percentage = currentPollLength / poll.userAnswers.length;
+    final BorderRadius borderRadiusResults = percentage >= 0.95
+        ? BorderRadius.circular(8)
+        : BorderRadius.only(
+            topLeft: Radius.circular(8),
+            bottomLeft: Radius.circular(8),
+          );
+    final Color selectedColor = pollColors[index % pollColors.length];
+
     return Container(
       child: Stack(
         alignment: Alignment.centerLeft,
         children: [
-          Positioned.fill(
+          FractionallySizedBox(
+            widthFactor: 1,
             child: Container(
-              width: MediaQuery.of(context).size.width * percentage,
+              height: 35.0,
               decoration: BoxDecoration(
                 borderRadius: BorderRadius.circular(8),
-                color: Theme.of(context).accentColor.withOpacity(0.25),
+                border: Border.all(width: 1.0, color: selectedColor),
+              ),
+            ),
+          ),
+          FractionallySizedBox(
+            widthFactor: percentage,
+            child: Container(
+              height: 35.0,
+              decoration: BoxDecoration(
+                borderRadius: borderRadiusResults,
+                color: selectedColor,
               ),
             ),
           ),
           Padding(
             padding: EdgeInsets.symmetric(
-              horizontal: ThemeSpaces.smallMargin,
+              horizontal: ThemeSpaces.mediumMargin,
               vertical: ThemeSpaces.smallMargin,
             ),
             child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
-                Text(
-                  "${(percentage * 100).toInt()}%",
-                  style: Theme.of(context)
-                      .textTheme
-                      .bodyText2
-                      .copyWith(fontWeight: FontWeight.bold),
+                Flexible(
+                  fit: FlexFit.loose,
+                  child: Text(
+                    option.text,
+                    softWrap: false,
+                    overflow: TextOverflow.fade,
+                    style: TextStyle(
+                      color: Theme.of(context).colorScheme.onPrimary,
+                    ),
+                  ),
                 ),
-                const SizedBox(width: 4),
-                Text(option.text),
+                RichText(
+                  text: TextSpan(
+                    style: TextStyle(
+                      color: Theme.of(context).colorScheme.onPrimary,
+                    ),
+                    children: <TextSpan>[
+                      TextSpan(
+                        text: '(${currentPollLength})',
+                        style: TextStyle(
+                          fontWeight: FontWeight.w300,
+                          fontSize: 14.0,
+                          color: Theme.of(context).colorScheme.onPrimary,
+                        ),
+                      ),
+                      TextSpan(
+                        text: ' ${(percentage * 100).toInt()}%',
+                        style: TextStyle(
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
               ],
             ),
           ),