+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Version 0.4.2
+## Changes
+- Add option to remove poll during post creation
+
+## Bug fixes
+- Fixed a bug that caused the funds to not be asked when needed
+- Fixed the visualization of poll results
+- Fixed a bug that caused a wrong visualization of a post details
+- Fixed user account edits to trim trailing white spaces on save
+
 # Version 0.4.1
 ## Changes
 - Added the proper error visualization inside the mnemonic recovery screen
@@ -5,7 +15,7 @@
 - Improved the account edit UX to make sure it's easier to change any information
 
 ## Bug fixes
-- Fixed the visualization of the posts' messages to make sure newlines are shown properly 
+- Fixed the visualization of the posts' messages to make sure newlines are shown properly
 - Fixed the visualization of the posts' times based on the devices location
 - Fixed the font visualization on iOS devices